fix: properly verify signature
Note that the signature HAS BEEN verified even
before this patch.
However, it relied on the fact that
the cast here `(JwsHeader) token.getHeader()` would
fail if the signature would not have been valid.
While it does indeed work, it is not good code, as relying
on casts for security is not good style.